style: redundant type cast.

+++ src/main/java/redis/persist/MetaChunkSegmentFlagSeq.java
@@ -152,7 +152,7 @@ public class MetaChunkSegmentFlagSeq {
         }
         try {
             for (int i = 0; i < initTimes; i++) {
-                raf.seek((long) i * BATCH_SIZE);
+                raf.seek( i * BATCH_SIZE);
                 raf.write(EMPTY_BYTES);
             }
             fillSegmentFlagInit(inMemoryCachedBytes);
