+++ src/main/java/redis/persist/StatKeyBucketLastUpdateCount.java
@@ -112,7 +112,7 @@ public class StatKeyBucketLastUpdateCount {
         }
         try {
             for (int i = 0; i < initTimes; i++) {
-                raf.seek((long) i * BATCH_SIZE);
+                raf.seek( i * BATCH_SIZE);
                 raf.write(EMPTY_BYTES);
             }
             Arrays.fill(inMemoryCachedBytes, (byte) 0);
